wx backend classes separated (except commented ones)
