+++ src/backends/wx/index.ts
@@ -1,0 +1,14 @@
+import { desktopSize, BaseElement } from "./BaseElement";
+import { AppElement } from "./AppElement";
+import { ButtonElement } from "./ButtonElement";
+import { ViewElement } from "./ViewElement";
+import { WindowElement } from "./WindowElement";
+
+export {
+  desktopSize,
+  BaseElement,
+  AppElement,
+  ButtonElement,
+  ViewElement, 
+  WindowElement
+}
